requirements.txt
opencv-python>=4.5.0
SpeechRecognition>=3.8.1
moviepy>=1.0.3
numpy>=1.19.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-google-genai>=0.0.5 
